requirements.txt
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
transformers==4.44.2
trafilatura==1.9.0
gradio==4.44.0
pandas==2.2.2
numpy==1.26.4
tqdm==4.66.4
requests==2.32.3
